refactor(test): consolidate extension-utils tests into boundary file
Move parseExtensionDir and removeFromExtensionsJson tests from
mock-based extension-utils.test.ts into the boundary test file,
converting removeFromExtensionsJson tests to use real filesystem.
